Strava-Ranking/StravaRanking/requirements.txt
Django==3.1.2
django-allauth==0.43.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pyjwt==1.7.1
python3-openid==3.2.0
requests==2.24.0
requests-oauthlib==1.3.0
